add popups edit profile anf edit avatar, data update with ref

+++ src/utils/api.js
@@ -60,21 +60,23 @@ class Api {
         })
         .then(res=>this._sendRequest(res))
     }
-    
-    addLike(id){
+
+    changeLikeStatus(id, isLiked){
+        if(isLiked){
+            
         return fetch(`${this._baseUrl}cards/${id}/likes`,{
             method: 'PUT',
             headers: this._headers
         })
             .then(res=>this._sendRequest(res))
-    }
-
-    deleteLike(id){
-    return fetch(`${this._baseUrl}cards/${id}/likes`,{
-        method: 'DELETE',
-        headers: this._headers
-    })
-        .then(res=>this._sendRequest(res))
+        }
+        else {
+            return fetch(`${this._baseUrl}cards/${id}/likes`,{
+                method: 'DELETE',
+                headers: this._headers
+            })
+                .then(res=>this._sendRequest(res))
+        }
     }
 
     changeAvatar(link){
